refactor: centralize shared type guard helpers
