+++ src/channels/msteams/utils.ts
@@ -5,6 +5,7 @@ import {
 
 export const MSTEAMS_CONVERSATION_REFERENCE_KEY =
   'msteams:conversation-reference';
+export { isRecord } from '../../utils/type-guards.js';
 export { normalizeValue };
 
 export function normalizeOptionalValue(value: unknown): string | null {
@@ -13,8 +14,4 @@ export function normalizeOptionalValue(value: unknown): string | null {
       ? normalizeValue(String(value))
       : '';
   return normalizeNullableTrimmedString(normalized);
-}
-
-export function isRecord(value: unknown): value is Record<string, unknown> {
-  return Boolean(value) && typeof value === 'object' && !Array.isArray(value);
 }
