+++ src/channels/voice/utils.ts
@@ -1,3 +1,1 @@
-export function isRecord(value: unknown): value is Record<string, unknown> {
-  return value !== null && typeof value === 'object' && !Array.isArray(value);
-}
+export { isRecord } from '../../utils/type-guards.js';
