+++ src/providers/utils.ts
@@ -1,12 +1,4 @@
-export function normalizeBaseUrl(baseUrl: string): string {
-  return String(baseUrl || '')
-    .trim()
-    .replace(/\/+$/g, '');
-}
-
-export function isRecord(value: unknown): value is Record<string, unknown> {
-  return !!value && typeof value === 'object' && !Array.isArray(value);
-}
+export { isRecord, normalizeBaseUrl } from '../utils/type-guards.js';
 
 export function readPositiveInteger(value: unknown): number | null {
   const parsed =
